Hide GameState and make GameEngine immutable
Exposing GameState to individual cards is problematic because its
functions care only about the manipulation of the game state; they don't
handle effect triggers and things like that.

Instead, we should treat it like a container, and have another layer
around it responsible for manipulating it and checking effects. This
upper layer should also be immutable, so that we can easily rollback /
dry-run actions without having to worry about state synchronization.

GameEngine was mostly immutable already; only the setup() function was
mutable, and it didn't need to be (it could just return the new
GameEngine). I've changed most function signatures in main() to
receive/return GameEngines instead of GameStates. Most of them still
reach for the internal data structure, but that's temporary; it will be
reworked in a future commit. GameState should only be accessible by
GameEngine.