Implement Gastly's Destiny Bond
I rewrote the whole effects implementation. Instead of trying to
enumerate all possible consequences, an effect is now an object that
implements a bunch of hooks, returning an AttackBuilder when
appropriate. This AttackBuilder then gets applied to the GameEngine.

Effects are now stored in state by name, and there's a mapping from
String -> CustomEffect in the Format object, similar to how there's a
String -> CardArchetype mapping.

AttackBuilder will probably be renamed to something else, as it's used
for more than attack implementations. It represents a state change
function.