Implement Clefairy's Metronome
the CardArchetype::attacks() function now returns all available attacks,
instead of the ones that can be performed. Metronome required this.

Thanks to this change, the carddb::Attacks struct became nothing but a
Vec<Action>, so I got rid of it. I also created an Attack struct that
contains the name and body of the attack, instead of relying on an
Action all the time.

Metronome ignores all costs, so I needed a way to know what part of the
attack is a cost. I added a OperationType to the Operation struct that
informs if it's a energy cost, another type of cost, another type of
requirement, or a normal operation.

Checking if an attack is enabled now only runs the requirement
operations. This reduces the cases where we get stuck in an infinite
loop.